+++ CHANGELOG.md
@@ -64,3 +64,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Modified Desc as parameter of fetch projects
 - Upload and download documents by document type
 - Fetch publications and other documents
+- Fetch project types